Fixed negative values in depth.

+++ src/fdf_count_colour_scale.c
@@ -6,11 +6,12 @@
 /*   By: jnivala <jnivala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/29 13:03:12 by jnivala           #+#    #+#             */
-/*   Updated: 2020/10/30 13:37:40 by jnivala          ###   ########.fr       */
+/*   Updated: 2020/10/31 11:35:45 by jnivala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "g42.h"
+#include <stdio.h>
 
 int		fdf_count_colour_scale(t_map *map)
 {
@@ -19,7 +20,7 @@ int		fdf_count_colour_scale(t_map *map)
 	double	c_dist;
 
 	i = 0;
-	minmax_dist = (double)((map->max_depth - map->min_depth));
+	minmax_dist = (double)map->max_depth - (double)map->min_depth;
 	while (i < map->pts)
 	{
 		c_dist = (double)(map->coord[i].y - map->min_depth);
